replica: extract query_state into own header
So it can be reused outside of replica/table.cc.
